voice_text_ai_assistant_gemini_flask/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
google-generativeai==0.8.3
openai-whisper==20231117
gTTS==2.5.3
pydub==0.25.1
